Fix for installing only one of certificates, sending user certificate password only when user certificate field is checked

+++ src/frontends/android/app/src/main/java/org/strongswan/android/security/LocalKeystore.java
@@ -204,10 +204,12 @@ public class LocalKeystore {
 
     public boolean removePkcs12AndCaCertificate(String certificateId) {
         try {
-            File[] certificateFiles = StrongSwanApplication.getContext().getFilesDir().listFiles();
-            for (File certificate : certificateFiles) {
-                if (certificate.getName().contains(certificateId)) {
-                    certificate.delete();
+            if(certificateId != null) {
+                File[] certificateFiles = StrongSwanApplication.getContext().getFilesDir().listFiles();
+                for (File certificate : certificateFiles) {
+                    if (certificate.getName().contains(certificateId)) {
+                        certificate.delete();
+                    }
                 }
             }
             return true;
